requirements.txt
PyYAML~=6.0.1
Cerberus~=1.3.4
pillow~=10.0.0
Flask~=2.3.2
werkzeug~=2.3.7
fake_rpi